meson: Provide build configuration summary
Report the enabled pipelines when configuring the meson build hierarchy.

This helps clarify what pipelines are used when building, and is useful
when looking back through build-traces to see what was actually
configured.

This can be expanded upon to report other options, and ideally will be
replaced by mesons summary() functionality when it is available.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>